SKALE-779 Restructure SSL commands
